+++ test/functional/mempool_accept.py
@@ -64,8 +64,10 @@ class MempoolAcceptanceTest(BitcoinTestFramework):
         self.log.info('Should not accept garbage to testmempoolaccept')
         assert_raises_rpc_error(-3, 'Expected type array, got string',
                                 lambda: node.testmempoolaccept(rawtxs='ff00baar'))
-        assert_raises_rpc_error(-8, 'Array must contain exactly one raw transaction for now',
-                                lambda: node.testmempoolaccept(rawtxs=['ff00baar', 'ff22']))
+        assert_raises_rpc_error(-8, 'Array must contain between 1 and 50 transactions.',
+                                lambda: node.testmempoolaccept(rawtxs=['ff22'] * 51))
+        assert_raises_rpc_error(-8, 'Array must contain between 1 and 50 transactions.',
+                                lambda: node.testmempoolaccept(rawtxs=[]))
         assert_raises_rpc_error(-22, 'TX decode failed',
                                 lambda: node.testmempoolaccept(rawtxs=['ff00baar']))